poll: Cast R_TIME_AS_MSECONDS to DWORD on the Win32 r_poll path
R_TIME_AS_MSECONDS produces an RClockTime (ruint64) but
WaitForMultipleObjects's timeout argument (and the local 't' that
holds it) is DWORD. Cast at the assignment to silence MSVC C4244.
Matches the same pattern applied to CreateTimerQueueTimer in
bae4ec3.

+++ rlib/rpoll.c
@@ -96,7 +96,7 @@ r_poll (RPoll * handles, ruint count, RClockTime timeout)
   if (timeout == R_CLOCK_TIME_INFINITE) {
     t = INFINITE;
   } else if (timeout != 0) {
-    t = R_TIME_AS_MSECONDS (timeout) + 1;
+    t = (DWORD)(R_TIME_AS_MSECONDS (timeout) + 1);
   } else {
     t = 0;
   }
